+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `snuffling.catalog` adding FDSN event query
 - Snuffler adding `trace_scale` to configuration
 - Added Burger elasticity parameters to cake and `gf.psgrn_pscmp`
+- Added Smithsonian Volcano database to `pyrocko.dataset`
+- Added Tile export to 3D printable formats
 
 ### Fixed
 - `seismosizer` STF timing of `RectangularSource`s.